initial support for generating PDF documentation

+++ tools/docs/gendocs.py
@@ -23,10 +23,30 @@ DOCS_DIRS = [
 
 def gendocs_xml(dir, doxyfile, version, doc_dir):
     doxycmd = GENERATE_XML.format(doxyfile, version, doc_dir)
-    subprocess.call((doxycmd), shell=True, cwd=dir)
+    return subprocess.call((doxycmd), shell=True, cwd=dir)
+
+def sphinx_build(outformat, srcdir='docs/', outdir=None, verbose=0):
+    if not outdir:
+        outdir = outformat + '/'
+
+    args = ['sphinx-build']
+    if verbose > 0:
+        args.append('-' + ('v' * verbose))
+
+    if outformat == 'pdf':
+        args.extend(['-b', 'latex'])
+
+    args.extend(['-D', 'outformat=%s' % outformat, srcdir, outdir])
+
+    if verbose > 0:
+        print(" ".join(args))
+
+    return subprocess.call(args)
 
 def main():
     parser = argparse.ArgumentParser()
+    parser.add_argument('--format', metavar='format', default='all',
+                        help='Which docs to generate: html, pdf, cargo, or all')
     parser.add_argument('--output', metavar='output', default='xml',
                         help='Specifies output directory for docs')
     parser.add_argument('--version', metavar='version', default='0.0.0',
@@ -34,21 +54,51 @@ def main():
 
     args = parser.parse_args()
 
+    if args.format not in ('html', 'pdf', 'cargo', 'all'):
+        parser.error('Unknown doc format: {}'.format(args.format))
+
     doc_tags = {}
 
     doc_dirs = [d for d in DOCS_DIRS if os.path.isdir(d)]
 
-    for dir in doc_dirs:
-        doc_dir = os.path.join(os.getcwd(), args.output, dir)
-        if not os.path.isdir(doc_dir):
-            os.makedirs(doc_dir)
-        gendocs_xml(dir, "docs/Doxyfile", args.version, doc_dir)
+    try:
+        for dir in doc_dirs:
+            doc_dir = os.path.join(os.getcwd(), args.output, dir)
+            if not os.path.isdir(doc_dir):
+                os.makedirs(doc_dir)
 
-    subprocess.call("sphinx-build docs/ html/", shell=True)
-    shutil.rmtree("./xml")
+            result = gendocs_xml(dir, "docs/Doxyfile", args.version, doc_dir)
+            if result != 0:
+                raise Exception(result, 'Error generating Doxygen in %s' % dir)
 
-    subprocess.call("cargo doc --all --no-deps", shell=True)
-    dir_util.copy_tree("target/doc", "html/rust-docs")
+        if args.format in ('html', 'all'):
+            result = sphinx_build('html')
+            if result != 0:
+                raise Exception(result, 'Error generating HTML')
+
+        if args.format in ('pdf', 'all'):
+            result = sphinx_build('pdf')
+            if result != 0:
+                raise Exception(result, 'Error generating LaTeX')
+
+            result = subprocess.call('make', cwd='pdf')
+            if result != 0:
+                raise Exception(result, 'Error transforming LaTeX to PDF')
+
+
+        if args.format in ('cargo', 'all'):
+            result = subprocess.call("cargo doc --all --no-deps", shell=True)
+            if result != 0:
+                raise Exception(result, 'Error generating cargo docs')
+
+            dir_util.copy_tree("target/doc", "html/rust-docs")
+    except Exception as e:
+        parser.exit(*e.args)
+    finally:
+        if os.path.isdir("./xml"):
+            shutil.rmtree("./xml")
+
+
 
 if __name__ == '__main__':
     main()
